restructure of imports & cleanup
