+++ src/instrument/plans/__init__.py
@@ -1,1 +1,5 @@
 """Bluesky plans."""
+
+from .sim_plans import sim_count_plan  # noqa: F401
+from .sim_plans import sim_print_plan  # noqa: F401
+from .sim_plans import sim_rel_scan_plan  # noqa: F401
