+++ src/instrument/plans/sim_plans.py
@@ -12,8 +12,9 @@ import logging
 
 from bluesky import plan_stubs as bps
 from bluesky import plans as bp
-from ophyd.sim import motor
-from ophyd.sim import noisy_det
+
+from ..devices import sim_det
+from ..devices import sim_motor
 
 logger = logging.getLogger(__name__)
 logger.bsdev(__file__)
@@ -24,8 +25,8 @@ DEFAULT_MD = {"title": "test run with simulator(s)"}
 def sim_count_plan(num: int = 1, imax: float = 10_000, md: dict = DEFAULT_MD):
     """Demonstrate the count() plan."""
     logger.debug("sim_count_plan()")
-    yield from bps.mv(noisy_det.Imax, imax)
-    yield from bp.count([noisy_det], num=num, md=md)
+    yield from bps.mv(sim_det.Imax, imax)
+    yield from bp.count([sim_det], num=num, md=md)
 
 
 def sim_print_plan():
@@ -33,7 +34,7 @@ def sim_print_plan():
     logger.debug("sim_print_plan()")
     yield from bps.null()
     print("sim_print_plan(): This is a test.")
-    print(f"sim_print_plan():  {motor.position=}  {noisy_det.read()=}.")
+    print(f"sim_print_plan():  {sim_motor.position=}  {sim_det.read()=}.")
 
 
 def sim_rel_scan_plan(
@@ -49,14 +50,14 @@ def sim_rel_scan_plan(
     logger.debug("sim_rel_scan_plan()")
     # fmt: off
     yield from bps.mv(
-        noisy_det.Imax, imax,
-        noisy_det.center, center,
-        noisy_det.sigma, sigma,
-        noisy_det.noise, noise,
+        sim_det.Imax, imax,
+        sim_det.center, center,
+        sim_det.sigma, sigma,
+        sim_det.noise, noise,
     )
     # fmt: on
-    print(f"sim_rel_scan_plan(): {motor.position=}.")
-    print(f"sim_rel_scan_plan(): {noisy_det.read()=}.")
-    print(f"sim_rel_scan_plan(): {noisy_det.read_configuration()=}.")
-    print(f"sim_rel_scan_plan(): {noisy_det.noise._enum_strs=}.")
-    yield from bp.rel_scan([noisy_det], motor, -span / 2, span / 2, num=num, md=md)
+    print(f"sim_rel_scan_plan(): {sim_motor.position=}.")
+    print(f"sim_rel_scan_plan(): {sim_det.read()=}.")
+    print(f"sim_rel_scan_plan(): {sim_det.read_configuration()=}.")
+    print(f"sim_rel_scan_plan(): {sim_det.noise._enum_strs=}.")
+    yield from bp.rel_scan([sim_det], sim_motor, -span / 2, span / 2, num=num, md=md)
